+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Support sampling, to reduce data volumes (#1)
 
 ### Fixed
+* Fix typo on meta tag names: s/broser/browser/
 * Stop capturing incomplete timing data from AJAX requests (#2)
 
 ## 0.3.0 / 2020-02-11