+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
+*no unreleased changes*
 
+## 0.4.0 / 2020-07-23
 ### Added
 * Support sampling, to reduce data volumes (#1)